init: adding compspec containment generation
The containment subsystem generates a graph for a cluster.
This is the minimal requirement for doing a feasibility check
or similar. We generate JGF now from the rlite, and I hope there
might be a better way (that better captures the shape of the graph).
I think this approach assumes a shape.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>